test: mark helpers, separate tests into proper files
